break out base view, tidy mod
